==> notebooks/03_load_postgre.ipynb <==

import pandas as pd
from pathlib import Path
from sqlalchemy import create_engine

# -----------------------------
# DB config
# -----------------------------
DB_USER = "postgres"
DB_PASS = "1998"
DB_HOST = "localhost"
DB_PORT = "5432"
DB_NAME = "rossmann"
DB_SCHEMA = "raw"

DB_URL = f"postgresql+psycopg2://{DB_USER}:{DB_PASS}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
engine = create_engine(DB_URL)

# -----------------------------
# Paths (absolute, stable)
# -----------------------------
PROJECT_ROOT = Path("C:/Users/User/Documents/Projects/dbt_rossmann_anayltics").resolve()
DATA_CLEAN_PATH = PROJECT_ROOT / "data_clean"

sales_path = DATA_CLEAN_PATH / "sales_clean.csv"
stores_path = DATA_CLEAN_PATH / "stores_clean.csv"



# -----------------------------
# Load STORES
# -----------------------------
if not stores_path.exists():
    raise FileNotFoundError(f"Missing file: {stores_path}")

print(f"\nLoading {stores_path.name} -> {DB_SCHEMA}.stores_raw")
stores_df = pd.read_csv(stores_path)
print(stores_df.head())
print(stores_df.dtypes)

stores_df.to_sql(
    name="stores_raw",
    con=engine,
    schema=DB_SCHEMA,
    if_exists="replace",
    index=False,
    chunksize=50_000,
    method="multi",
)
print(f"Loaded {len(stores_df):,} rows into {DB_SCHEMA}.stores_raw")

print("\nDone.")



# -----------------------------
# Load SALES
# -----------------------------
if not sales_path.exists():
    raise FileNotFoundError(f"Missing file: {sales_path}")

print(f"\nLoading {sales_path.name} -> {DB_SCHEMA}.sales_raw")
sales_df = pd.read_csv(sales_path)
print(sales_df.head())
print(sales_df.dtypes)

sales_df.to_sql(
    name="sales_raw",
    con=engine,
    schema=DB_SCHEMA,
    if_exists="replace",   # avoids duplicates while developing
    index=False,
    chunksize=100_000,
    method="multi",
)
print(f"Loaded {len(sales_df):,} rows into {DB_SCHEMA}.sales_raw")
